Reverting FakeDetector to macros.py to test hutch-python dev workflow.
